Move package slack/message/handler
